display fahrenheit or celsius as preference.
