+++ app/src/main/res/values/strings.xml
@@ -1,4 +1,5 @@
 <resources>
+
     <string name="app_name" translatable="false">Sunshine</string>
     <string name="hello_world" translatable="false">Hello world!</string>
     <string name="action_settings">Settings</string>
@@ -9,4 +10,17 @@
     <string name="pref_location_key" translatable="false">location</string>
     <string name="pref_location_label">Location</string>
     <string name="pref_location_default" translatable="false">Guangzhou</string>
+
+    <string name="pref_unit_default" translatable="false">metric</string>
+    <string name="pref_unit_label">Temperature Unit</string>
+    <string name="pref_unit_key" translatable="false">unit</string>
+    <string-array name="pref_unit_entries">
+        <item>Celsius</item>
+        <item>Fahrenheit</item>
+    </string-array>
+    <string-array name="pref_unit_values">
+        <item>metric</item>
+        <item>imperial</item>
+    </string-array>
+
 </resources>
