+++ app/src/main/res/xml/pref_general.xml
@@ -8,4 +8,12 @@
         android:singleLine="true"
         android:title="@string/pref_location_label" />
 
+    <ListPreference
+        android:defaultValue="@string/pref_unit_default"
+        android:dialogTitle="@string/pref_unit_label"
+        android:entries="@array/pref_unit_entries"
+        android:entryValues="@array/pref_unit_values"
+        android:key="@string/pref_unit_key"
+        android:title="@string/pref_unit_label" />
+
 </PreferenceScreen>
